Move the User's methods; Remove the data.go file
Update go.mod

Update README.md

Add .gitignore
