+++ README.md
@@ -1,4 +1,4 @@
-# Практичні 2-го етапу курсу з Go
+# Рішення практичних завдань (етап 2.1)
 
 |  # | Короткий опис                                                  | Умова | Рішення |
 |---:|:---------------------------------------------------------------|-------|---------|
@@ -7,8 +7,8 @@
 |  3 | Відсортувати користувачів за сумою середн. віку читача кожної книжки ... | [#5](https://github.com/grescher/goc-s2-psets/issues/5) | [p3.0](https://github.com/grescher/goc-s2-psets/releases/tag/p3.0) |
 |  4 | Типізовані константи: додати поле "тип користувача"...         | [#6](https://github.com/grescher/goc-s2-psets/issues/6) | [p4.0](https://github.com/grescher/goc-s2-psets/releases/tag/p4.0) |
 |  5 | Декодування даних з бінарного буферу                           | [#8](https://github.com/grescher/goc-s2-psets/issues/8) | [p5.0](https://github.com/grescher/goc-s2-psets/releases/tag/p5.0) |
-|  6 | Спільний індекс — стан активності користувачі                  | [#10](https://github.com/grescher/goc-s2-psets/issues/10) | |
-|  7 | Збереження у файл з допомогою елементарних команд              | [#12](https://github.com/grescher/goc-s2-psets/issues/12) | |
-|  8 | Розпізнавання виразів з цифрами                                | [#13](https://github.com/grescher/goc-s2-psets/issues/13) | |
+|  6 | Спільний індекс — стан активності користувачі                  | [#10](https://github.com/grescher/goc-s2-psets/issues/10) | [p6.0](https://github.com/grescher/goc-s2-psets/releases/tag/p6.0) |
+|  7 | Збереження у файл. Елементарний інтерпретатор на `Scanf`       | [#12](https://github.com/grescher/goc-s2-psets/issues/12) | |
+|  8 | Розпізнавання виразів з цифрами. Буферизований ввод.           | [#13](https://github.com/grescher/goc-s2-psets/issues/13) | |
 |  9 | Пересилання даних між клієнтом і сервером                      | [#14](https://github.com/grescher/goc-s2-psets/issues/14) | |
-| 10 | Зробити http-сервер, який отримує дані з БД (API) ...          | [#15](https://github.com/grescher/goc-s2-psets/issues/15) | |
+| 10 | http-сервер, який отримує дані з БД (API)                      | [#15](https://github.com/grescher/goc-s2-psets/issues/15) | |
