+++ go.mod
@@ -1,5 +1,5 @@
 module practice
 
-go 1.20
+go 1.21
 
 require golang.org/x/exp v0.0.0-20230711153332-06a737ee72cb
